Modify the optional switches for sanity-check code in function `probe`:
 1. Remove optional switch `verify_ffprobe_found` & its corresponding sanity-check code.
 2. Modify the behaviour of the local-file-exists sanity-check code; and rename its optional switch `verify_mediafile_found` to `verify_local_mediafile`.

Details:  There were 2 main changes in this commit, both relating to the optional boolean switch parameters of function `ffprobe3.probe` & the sanity-check code that these switches control:

Change 1:  In function `ffprobe3.probe`, I removed the first sanity-check code block (controlled by optional switch `verify_ffprobe_found`; entered when `verify_ffprobe_found=True`, the default), which used function `subprocess.check_call` to call `[ffprobe_cmd, '-h']` to verify that the `ffprobe` command can be found in the `$PATH`.

I realised that by handling exception `FileNotFoundError` specifically at the later `subprocess.Popen` function call (the subprocess call which initiates the actual media file probe using `ffprobe`), I could unambiguously detect when the specified `ffprobe` executable cannot be found.  Thus, there was no longer any need for a prior check using an additional subprocess call.  And if there is no additional subprocess call beforehand, then there's no need for an optional boolean switch to disable it.

So, I removed the optional switch `verify_ffprobe_found` in the parameters of function `probe`; I moved the specific handling of exception `FileNotFoundError` down to the exception handling for `subprocess.Popen`; and I deleted the whole code block controlled by `verify_ffprobe_found`.

Change 2:  In function `ffprobe3.probe`, I modified the behaviour of the second sanity-check code block (controlled by optional switch `verify_mediafile_found`; entered when `verify_mediafile_found=True`, the default) so that it will *only* perform the local-file-exists check when `verify_mediafile_found=True` **AND** the specified media source is *obviously* a local file.  (And I renamed this optional switch from `verify_mediafile_found` to `verify_local_mediafile`.)

Previously, the concept behind this sanity check was simple:  "Does the specified media source exist as a local file?"

I reasoned that if the specified media source *was* a local file, then it was reasonable to check whether it existed (at the cost of an extra filesystem access, which is why you might possibly want to disable the check).

And if the specified media source was *not* a local file... was that an accident?  Accessing remote URIs is expensive.  If you really *intended* to specify a remote media URI, then you could simply disable the check.

But now I've decided that it's probably highly unlikely for client code to specify a remote media URI *by accident*.  So I've changed the behaviour of the sanity-check code block so that the local-file-exists check *only* occurs when the specified media source is obviously a local file (ignoring the issue of NFS filesystems, etc.).

The new code considers a specified media source to be a local file if it is **NOT** preceded by a "protocol" (as `ffprobe` calls it) / "URI scheme" (as the Internet calls it) such as `"http://"`.

I note that `ffprobe` does actually provide a list of the "protocols" that it supports, on request (using `ffprobe -protocols`).  But rather than hard-coding this list of supported protocols -- or adding an extra subprocess call, just to query the current list of protocols -- I've instead just written a simple regex to detect a prefix that looks like a URI scheme (see the global variable `_URI_SCHEME`).  No exceptions made for `"file://"`.

And because the behaviour of the sanity check code has changed subtly, I renamed the optional switch parameter from `verify_mediafile_found` to `verify_local_mediafile`.