Make class `ParsedJson` inherit from `Mapping` ABC.
So that class `ParsedJson` implements the `Mapping` abstract base class, it now has the following new methods:

- method `__contains__(key)`
- method `__eq__(other)`
- method `__getitem__(key)`
- method `__iter__()`
- method `__len__()`

Also added 2 new class API introspection methods to class `ParsedJson`:

- method `get_attr_names()`
- method `get_getter_names()`

Also wrote *lots* more docstring documentation for class `ParsedJson`, as the base class of all the other classes (and the real workhorse of the classes).

In particular, this `ParsedJson` docstring documentation now emphasises that an instance of this class will reference (and share) the original `dict` instance of parsed JSON that was supplied to its constructor.  It will *not copy* the argument `dict` items into a new `dict` owned exclusively by this instance.  As a result, for a given `ffprobe` output, all instances of derived classes of `ParsedJson` will *share the same original tree* of `dict` instances that was returned by `json.loads`; they will each reference a different subtree of the shared tree.

I also wrote a lot of example usage of the `ParsedJson` class API within its class & method docstrings (especially for `__repr__()`, `get_attr_names()`, `get_getter_names()`, and `keys()`).

And finally, as extremely minor changes:

- I renamed `ParsedJson` method `get_duration_as_HH_MM_SS_ss` to the easier-to-remember `get_duration_as_human`.
- I changed `FFformat.bit_rate_kbps`, `FFaudioStream.bit_rate_kbps`, and `FFvideoStream.bit_rate_kbps`, from type `int` to `float`.
- To initialise the values of attributes in derived classes of `ParsedJson`, I use `.get_as_int` or `.get_as_float` rather than `.get` whenever possible, to reduce the number of manual type conversions necessary by client code.