Init commit of new test module `tests/test_ffprobe3.py`.
This new test module is for the recently-new `ffprobe3/ffprobe3.py` module in this recently-new fork of the repo:  It tests the new API of classes & functions in `ffprobe3`.  The tests are a series of `assert` statements about the values returned by the API: crude but comprehensive.  (All of them pass right now.)

This new test module will replace the previous test module `tests/ffprobe_test.py`, which is for the original pre-fork `ffprobe` module (which no longer exists).  (It's worth noting that the previous test module didn't `assert` anything; instead, it merely printed values.  So it was purely a test of the existence of the API classes, not a test of the values returned by the API.)

As an additional minor change, I also added an extra attribute `num_frames` to the stream classes `FFaudioStream` & `FFvideoStream`.