Delete old test module `tests/ffprobe_test.py`.
This old test module `tests/ffprobe_test.py` is for the original pre-fork `ffprobe` module (which no longer exists in this repo).

(It's worth noting that this old test module didn't `assert` anything; instead, it merely printed values.  So it was purely a test of the existence of the API classes, not a test of the values returned by the API.)

This old test module has been replaced by new test module `tests/test_ffprobe3.py`.  The new test module is for the recently-new `ffprobe3/ffprobe3.py` module in this recently-new fork of the repo:  It tests the new API of classes & functions in `ffprobe3`.