Update `setup.py` (setuptools) for package `ffprobe3-python3`.
I updated all the fields that seemed like they needed to change:

- `name`
- `version`
- `description`
- `author`
- `author_email`
- `maintainer`
- `maintainer_email`
- `url`
- `packages`
- `keywords`
- `classifiers`

I haven't read up on the `setup.py` format -- I just modified the existing file.  So here's hoping this works!

As a minor detail of note:  I changed the string in field `description` so that it's just a single-line string (ie, no embedded newlines).  Previously it was a multi-line string using triple-quotes.  I wonder if this change would fix this reported issue #18 for `ffprobe-python`?
 https://github.com/gbstack/ffprobe-python/issues/18