Fix an occasional error in the human-readable duration string.
The error was caused by an incorrect format string in method `ParsedJson.get_duration_as_human`.  Thanks to James Baucom for spotting the error in the format string and suggesting a fix!

As James Baucom explained:
"""
I noticed that the duration_human time format doesn't zero pad seconds if they are less than 10. I think that instead of "%02d:%02d:%02.2f" the format should be "%02d:%02d:%05.2f". The first number for a float represents the entire string, and not just the integer part.
"""